pyproject.toml
[project]
name = "realtime-dashboard"
version = "0.1.0"
description = ""
authors = [
    {name = "Rafael Souza",email = "rafasansou1@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "streamlit (>=1.46.1,<2.0.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "streamlit-autorefresh (>=1.0.1,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
